bug fixes #3

+++ core/src/it/simone/davide/cardtd/classes/Damageable.java
@@ -2,5 +2,5 @@ package it.simone.davide.cardtd.classes;
 
 public interface Damageable {
 
-    public void damage(int damage);
+    public void damage(float damage);
 }
